#2 Move classes to the separate files
