+++ reliability/part.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+__all__ = ['Part']
+
+
+@dataclass(frozen=True)
+class Part:
+    type_: int
+    λ: float
